reorg server structure, get working server main test inputs
