+++ skiday/main.py
@@ -1,0 +1,3 @@
+from server import services
+
+services.update_resort_data('cannon')
